Make the library compatible with Lilypond 2.23.
harmony-rules-tests.scm uses ‘local-eval’ which is not available by
default with Guile 2.2; use a conditional ‘use-modules’.

harmony-rules.scm & harmony-rules-tests.scm used ‘print’ that was
removed from Lilypond in 2.23.

harmony-rules.scm used ‘<Prob>’ to identify the class of "music"
objects in Lilypond. This class is not longer visible in 2.23.

harmony-rules.scm identified bars with ‘ContextSpeccedMusic’ or type
‘Timing’ with property ‘whichBar’. Lilypond 2.23 now uses ‘BarEvent’
with property ‘bar-type’ instead.

harmony-rules.scm somehow used a ‘case’ with ‘=>’ incorrectly. For
some reason Guile 1.8 did not complained about it (it should have) but
Guile 2.2 refuse it.